Monitor colours should change as soon as the colour button is set
In order to change a monitor's colour in the visualisation, you
need to select a monitor on the Devices tab, set the colour via the
Colour button at the bottom, then use the Change button and OK the
dialog that pops up.

The way the UI is made, it suggests the colour should update
as soon as the Colour dialog is OK'd, but it looks like it never
did - this has now been fixed.

Reported by alwayseverything@mail.com in #13803 /
https://bugzilla.xfce.org/show_bug.cgi?id=13803